requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
aerich==0.5.8
    # via -r requirements.in
aiohttp==3.8.1
    # via -r requirements.in
aiosignal==1.2.0
    # via aiohttp
aiosqlite==0.17.0
    # via tortoise-orm
anyio==3.3.4
    # via starlette
asgiref==3.4.1
    # via uvicorn
async-timeout==4.0.1
    # via aiohttp
asyncpg==0.25.0
    # via tortoise-orm
attrs==21.2.0
    # via aiohttp
bcrypt==3.2.0
    # via -r requirements.in
beautifulsoup4==4.10.0
    # via -r requirements.in
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via bcrypt
charset-normalizer==2.0.7
    # via
    #   aiohttp
    #   requests
click==8.0.3
    # via
    #   aerich
    #   uvicorn
ddlparse==1.10.0
    # via aerich
dictdiffer==0.9.0
    # via aerich
fastapi==0.70.0
    # via
    #   -r requirements.in
    #   fastapi-jwt-auth
fastapi-jwt-auth==0.5.0
    # via -r requirements.in
finnhub-python==2.4.5
    # via -r requirements.in
frozenlist==1.2.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.12.0
    # via uvicorn
httptools==0.2.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   requests
    #   yarl
iso8601==0.1.16
    # via tortoise-orm
multidict==5.2.0
    # via
    #   aiohttp
    #   yarl
psycopg2-binary==2.9.2
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic==1.8.2
    # via
    #   aerich
    #   fastapi
pyjwt==1.7.1
    # via fastapi-jwt-auth
pyparsing==3.0.6
    # via ddlparse
pypika-tortoise==0.1.1
    # via tortoise-orm
python-dotenv==0.19.2
    # via
    #   -r requirements.in
    #   uvicorn
pytz==2021.3
    # via tortoise-orm
pyyaml==6.0
    # via uvicorn
requests==2.26.0
    # via finnhub-python
six==1.16.0
    # via bcrypt
sniffio==1.2.0
    # via anyio
soupsieve==2.3.1
    # via beautifulsoup4
starlette==0.16.0
    # via fastapi
tortoise-orm[asyncpg]==0.17.8
    # via
    #   -r requirements.in
    #   aerich
typing-extensions==4.0.0
    # via
    #   aiosqlite
    #   async-timeout
    #   pydantic
urllib3==1.26.7
    # via requests
uvicorn[standard]==0.15.0
    # via -r requirements.in
uvloop==0.16.0
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==10.1
    # via uvicorn
yarl==1.7.2
    # via aiohttp
